Fix bug in table caption not working for main part of doc. Continue work on refactoring
