+++ src/paradoc/formatting/__init__.py
@@ -1,0 +1,3 @@
+from .concepts import Formatting, TableFormat
+
+__all__ = [Formatting, TableFormat]
