+++ tests/test_documentation.py
@@ -65,7 +65,7 @@ class MyCompositions(unittest.TestCase):
         document.add_page_break()
         document.add_paragraph("My Appendix", style="Appendix")
         document.add_paragraph("A subheading of my app", style="Appendix X.1")
-        document.add_paragraph("A subheading of a subheading in my app", style="Appendix X.1.1")
+        document.add_paragraph("A subheading of a subheading in my app", style="Appendix X.2")
         os.makedirs(test_dir, exist_ok=True)
         document.save(test_dir / "my_demo.docx")
